refactor(cardano-services): add search stake pool builder
move DbSyncStakePoolSearchProvider methods to StakePoolSearchBuilder
